fix(web): wire desktop topbar workspace switching through last-route restore (TASK-754 follow-up)
The TASK-754 restore logic only fired from WorkspaceSwitcher.svelte
(used on mobile). On DESKTOP, the workspace switcher is the topbar's
horizontal workspace icon list, which used plain `<a href>` links to
`/{owner}/{slug}` — bypassing restore entirely and silently
overwriting the workspace's saved deep route on every left-click.

Symptom (reported by user): "navigate to a deep page → storage updates
to that page → navigate to another workspace → saved value sticks →
click back via topbar → lands on dashboard, and the saved value gets
overwritten back to dashboard."

Fix:
- Extract the validation+pickup logic into a pure helper at
  `web/src/lib/utils/workspace-route.ts` (`workspaceRestoreTarget`).
- WorkspaceSwitcher.svelte's `select()` now delegates to the helper
  (no behavior change on mobile).
- TopBar.svelte intercepts plain left-click on each workspace `<a>` to
  goto the restore target. `href=` stays pointed at the dashboard so
  modifier-clicks (cmd/ctrl/shift/alt) and middle-click still open a
  fresh dashboard in a new tab.

Other workspace nav surfaces are left alone on purpose:
- Sidebar Dashboard nav item, mobile-header workspace name, and
  /console workspace cards are not "switchers" — semantically they're
  Home/breadcrumb/picker navigation that should always land on the
  dashboard.

Parent: IDEA-753.
