+++ web/src/lib/components/layout/WorkspaceSwitcher.svelte
@@ -3,6 +3,7 @@
 	import { workspaceStore } from '$lib/stores/workspace.svelte';
 	import { uiStore } from '$lib/stores/ui.svelte';
 	import BottomSheet from '$lib/components/common/BottomSheet.svelte';
+	import { workspaceRestoreTarget } from '$lib/utils/workspace-route';
 
 	interface Props {
 		/**
@@ -60,45 +61,9 @@
 		// workspace links did this on click, so preserve the behavior now
 		// that the switcher is the mobile nav entry point.
 		uiStore.onNavigate();
-
-		// Restore the last-visited route in this workspace if we have one
-		// in localStorage (written by the workspace +layout on every nav).
-		// Implements IDEA-753 / TASK-754.
-		//
-		// The saved value is treated as untrusted input — it could be
-		// stale, corrupt, or crafted. We canonicalize through `URL` and
-		// require:
-		//
-		//  - same origin (rejects scheme/`//host` smuggling),
-		//  - workspace prefix on the *normalized* pathname,
-		//  - no traversal segments or percent-encoded chars in the path
-		//    (the app never generates either; rejecting them blocks
-		//    `%2e%2e/...` style escapes that could pass a naive prefix
-		//    check but resolve elsewhere via `goto()` normalization).
-		//
-		// Anything failing the gauntlet falls back to the dashboard.
-		const fallback = `/${ws.owner_username}/${ws.slug}`;
-		let target = fallback;
-		try {
-			const saved = localStorage.getItem(`pad-last-route-${ws.slug}`);
-			if (saved) {
-				const normalized = new URL(saved, window.location.origin);
-				const path = normalized.pathname;
-				const sameOrigin = normalized.origin === window.location.origin;
-				const inWorkspace = path === fallback || path.startsWith(fallback + '/');
-				const cleanPath =
-					!path.includes('/..') &&
-					!path.includes('/./') &&
-					!path.includes('//') &&
-					!/%[0-9a-fA-F]{2}/.test(path);
-				if (sameOrigin && inWorkspace && cleanPath) {
-					target = normalized.pathname + normalized.search + normalized.hash;
-				}
-			}
-		} catch {
-			// localStorage unavailable or URL parse failed; fall through to dashboard.
-		}
-		goto(target);
+		// Restore the last-visited route in this workspace if cached
+		// (TASK-754). Validation lives in workspaceRestoreTarget.
+		goto(workspaceRestoreTarget(ws));
 	}
 
 	function openCreateModal() {
